requirements.txt
aniso8601==7.0.0
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
Django==3.1.3
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
promise==2.3
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pytz==2020.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
